+++ R/beeca-package.R
@@ -5,3 +5,6 @@
 #' @importFrom lifecycle deprecated
 ## usethis namespace: end
 NULL
+
+## declare global variables
+utils::globalVariables(c("estimate", "std.error", "statistic"))