+++ NEWS.md
@@ -1,4 +1,7 @@
-# beeca (development version)
+# beeca 0.1.4
+
+- Add `tidy_beeca()` function
+- Add Readme.Rmd as a Rmarkdown file + github action
 
 # beeca 0.1.3